added querys

+++ install.sh
@@ -63,19 +63,36 @@ echo -e "${YELLOW}Enabling GDM service...${NC}"
 sudo systemctl enable gdm.service
 
 # install gnome themes and icons...
-sudo pacman -S gnome-tweaks --noconfirm
-
-mkdir ~/.themes
-mkdir ~/.icons
-cp -r ./systemconfig/themes/* ~/.themes/
-cp -r ./systemconfig/icons/* ~/.icons/
+echo -e "${YELLOW}Do you want to install gnome tweaks? (y/n)${NC}"
+    read -r -p "Answer: " response
+    case "$response" in
+        ""|[yY][eE][sS]|[yY])
+            sudo pacman -S gnome-tweaks --noconfirm
+            mkdir ~/.themes
+            mkdir ~/.icons
+            cp -r ./systemconfig/themes/* ~/.themes/
+            cp -r ./systemconfig/icons/* ~/.icons/
+            ;;
+        *)
+            echo -e "${YELLOW}Skipping...${NC}"
+            ;;
+    esac
 
 # setup openrgb
-bash ./dotfiles/openrgbconf/openrgb.sh
+echo -e "${YELLOW}Do you want to install openrgb? (y/n)${NC}"
+    read -r -p "Answer: " response
+    case "$response" in
+        ""|[yY][eE][sS]|[yY])
+            bash ./dotfiles/openrgbconf/openrgb.sh
+            ;;
+        *)
+            echo -e "${YELLOW}Skipping...${NC}"
+            ;;
+    esac
 
 # Ask for Reboot
 echo -e "${GREEN}Installation complete! GNOME is set up and running.${NC}"
-echo -e "${GREEN}If you want to install the gnome extensions, you will need to start the /systemconfig/gnomeextensions.sh, after you log into gnome${NC}"
+echo -e "${YELLOW}If you want to install the gnome extensions, you will need to start the /systemconfig/gnomeextensions.sh, after you log into gnome${NC}"
 echo -e "${YELLOW}Do you want to reboot? (y/n)${NC}"
     read -r -p "Answer: " response
     case "$response" in
